+++ include/kfs/console.h
@@ -5,7 +5,7 @@
  * 既存実装(drivers/video/terminal.c)から宣言のみ抽出。機能追加なし。
  */
 
-#include <linux/vga.h>
+#include <video/vga.h>
 #include <stddef.h>
 #include <stdint.h>
