+++ include/kfs/string.h
